requirements.txt
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-openai
langchain-anthropic
textual
python-dotenv
duckduckgo-search
google-generativeai
